# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Ohio Casino API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /games:
    get:
      operationId: listGames
      tags: [games]
      summary: List all casino games
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: featured
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of games
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Game"
              example:
                - id: 1
                  name: "Diamond Rush"
                  category: "slots"
                  provider: "NetEnt"
                  rtp: 0.96
                  minBet: 100
                  maxBet: 100000
                  imageUrl: "/games/diamond-rush.jpg"
                  isFeatured: true
                  isHot: false
                  isNew: true
                  description: "A dazzling slot adventure"
                  jackpotAmount: 5000000
  /games/{id}:
    get:
      operationId: getGame
      tags: [games]
      summary: Get game details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Game details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
              example:
                id: 1
                name: "Diamond Rush"
                category: "slots"
                provider: "NetEnt"
                rtp: 0.96
                minBet: 100
                maxBet: 100000
                imageUrl: "/games/diamond-rush.jpg"
                isFeatured: true
                isHot: false
                isNew: true
                description: "A dazzling slot adventure"
                jackpotAmount: 5000000
        "404":
          description: Not found
...
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: "#/components/parameters/AuthorizationSessionHeader"
        - $ref: "#/components/parameters/SessionCookie"
      responses:
        "200":
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserEnvelope"
              example:
                user:
                  id: "user_123"
                  email: "player@example.com"
                  firstName: "John"
                  lastName: "Doe"
                  profileImageUrl: "https://example.com/avatar.jpg"
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        "302":
          description: Redirect to the OIDC authorization endpoint.
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: iss
          in: query
          required: false
          schema:
            type: string
            format: uri
      responses:
        "302":
          description: Redirect to the returnTo path on success.
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: "#/components/parameters/AuthorizationSessionHeader"
        - $ref: "#/components/parameters/SessionCookie"
      responses:
        "302":
          description: Redirect to the OIDC provider logout URL.
  /mobile-auth/token-exchange:
    post:
      tags: [Auth]
      operationId: exchangeMobileAuthorizationCode
      summary: Exchange a mobile OIDC code for a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MobileTokenExchangeRequest"
      responses:
        "200":
          description: Session created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MobileTokenExchangeSuccess"
        "400":
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Token did not contain usable claims.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Token exchange failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /mobile-auth/logout:
    post:
      tags: [Auth]
      operationId: logoutMobileSession
      summary: Delete a mobile session token
      parameters:
        - $ref: "#/components/parameters/AuthorizationSessionHeader"
      responses:
        "200":
          description: Logout processed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogoutSuccess"
...
  /wallet:
    get:
      operationId: getWallet
      tags: [wallet]
      summary: Get the authenticated user's wallet
      parameters:
        - $ref: "#/components/parameters/AuthorizationSessionHeader"
        - $ref: "#/components/parameters/SessionCookie"
      responses:
        "200":
          description: Wallet details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
              example:
                id: 1
                userId: 987
                balance: 9500
                currency: "USD"
                createdAt: "2025-01-15T08:00:00Z"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error: "Unauthorized"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error: "Internal server error"
  /wallet/history:
    get:
      operationId: getWalletHistory
      tags: [wallet]
      summary: Get wallet transaction history
      parameters:
        - $ref: "#/components/parameters/AuthorizationSessionHeader"
        - $ref: "#/components/parameters/SessionCookie"
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: offset
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Transaction history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
              example:
                - id: 1
                  walletId: 1
                  userId: 987
                  type: "bet"
                  amount: -1000
                  balanceBefore: 10500
                  balanceAfter: 9500
                  status: "completed"
                  referenceId: "round_12345"
                  description: "Bet on slots"
                  createdAt: "2025-06-23T10:00:00Z"
                - id: 2
                  walletId: 1
                  userId: 987
                  type: "win"
                  amount: 5000
                  balanceBefore: 9500
                  balanceAfter: 14500
                  status: "completed"
                  referenceId: "round_12345"
                  description: "Win on slots"
                  createdAt: "2025-06-23T10:01:00Z"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error: "Unauthorized"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              example:
                error: "Internal server error"
...
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: [string, "null"]
          format: email
        firstName:
          type: [string, "null"]
        lastName:
          type: [string, "null"]
        profileImageUrl:
          type: [string, "null"]
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: "#/components/schemas/AuthUser"
            - type: "null"
    MobileTokenExchangeRequest:
      type: object
      required: [code, code_verifier, redirect_uri, state]
      properties:
        code:
          type: string
          minLength: 1
        code_verifier:
          type: string
          minLength: 1
        redirect_uri:
          type: string
          minLength: 1
          format: uri
        state:
          type: string
          minLength: 1
        nonce:
          type: string
          minLength: 1
    MobileTokenExchangeSuccess:
      type: object
      required: [token]
      properties:
        token:
          type: string
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Game:
      type: object
      required:
        [id, name, category, provider, rtp, imageUrl, isFeatured, isHot, isNew]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
          description: "slots | blackjack | roulette | poker | sports | live"
        provider:
          type: string
        rtp:
          type: number
          description: Return to player percentage
        minBet:
          type: number
        maxBet:
          type: number
        imageUrl:
          type: string
        isFeatured:
          type: boolean
        isHot:
          type: boolean
        isNew:
          type: boolean
        description:
          type: ["string", "null"]
        jackpotAmount:
          type: ["number", "null"]
...
    Wallet:
      type: object
      required: [id, userId, balance, currency, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        balance:
          type: integer
          description: Balance in cents
        currency:
          type: string
        createdAt:
          type: string
...
    Transaction:
      type: object
      required:
        [
          id,
          walletId,
          userId,
          type,
          amount,
          balanceBefore,
          balanceAfter,
          status,
          createdAt,
        ]
      properties:
        id:
          type: integer
          example: 1
        walletId:
          type: integer
          example: 1
        userId:
          type: integer
          example: 987
        type:
          type: string
          example: "bet"
        amount:
          type: integer
          description: Amount in cents
          example: -1000
        balanceBefore:
          type: integer
          description: Balance before in cents
          example: 10500
        balanceAfter:
          type: integer
          description: Balance after in cents
          example: 9500
        status:
          type: string
          example: "completed"
        referenceId:
          type: ["string", "null"]
          example: "round_12345"
        description:
          type: ["string", "null"]
          example: "Bet on slots"
        createdAt:
          type: string
          example: "2025-06-23T10:00:00Z"